refactor: updated int to more accurate types

+++ src/entities/whip.h
@@ -1,7 +1,7 @@
 #ifndef _WHIP_H
 #define _WHIP_H
 
-Entity *whipSpawn(int x, int y, bool facingRight);
+Entity *whipSpawn(uint16_t x, uint16_t y, bool facingRight);
 
 void whipUpdate(struct Entity *e, uint8_t tileCollisions);
